refactor(session-replay): isolate policy wildcard matching
